rest_backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
channels==4.0.0
daphne==4.1.0
channels-redis==4.2.0
pyserial==3.5
requests==2.31.0
websockets==12.0
